Fix type warnings & memory leak
Making sure that types warnings are all acknowledged and dealt with.  Also, fixed a memory leak in GTIndexEntry.

+++ Classes/GTCommit.m
@@ -75,9 +75,9 @@
 						 parents:(NSArray *)theParents 
 						   error:(NSError **)error {
 	
-	int count = theParents ? theParents.count : 0;
+	NSUInteger count = theParents ? theParents.count : 0;
 	git_commit const *parentCommits[count];
-	for(int i = 0; i < count; i++){
+	for (NSUInteger i = 0; i < count; i++){
 		parentCommits[i] = ((GTCommit *)[theParents objectAtIndex:i]).commit;
 	}
 	
@@ -90,7 +90,7 @@
 									   committerSig.sig, 
 									   newMessage ? [newMessage UTF8String] : "", 
 									   theTree.tree, 
-									   count, 
+									   (int)count, 
 									   parentCommits);
 	if(gitError != GIT_SUCCESS) {
 		if(error != NULL)
